feat: add localApi
